refactor: HTTP 응답 관련 클래스들을 response 디렉토리로 재구성
- src/http/serializer 디렉토리를 src/http/response로 이름 변경
- Factory와 Serializer를 src/http/response 디렉토리로 이동
- 응답 생성 및 전송 관련 로직의 응집도 향상
