+++ src/handler/builder/ErrorBuilder.cpp
@@ -1,7 +1,7 @@
 // ErrorBuilder.cpp
 #include "ErrorBuilder.hpp"
 
-#include "../../http/ResponseFactory.hpp"
+#include "../../http/response/Factory.hpp"
 
 using namespace handler::builder;
 
@@ -17,7 +17,7 @@ std::string ErrorBuilder::joinAllowMethods(const std::vector<std::string>& metho
 
 http::Packet ErrorBuilder::build(const router::RouteDecision& decision, const http::Packet&,
 								 const config::Config& config) const {
-	http::Packet response = http::ResponseFactory::createError(
+	http::Packet response = http::response::Factory::createError(
 		decision.status, &config, http::StatusCode::to_reasonPhrase(decision.status),
 		http::ContentType::to_string(http::ContentType::CONTENT_TEXT_PLAIN));
 	std::string allow = joinAllowMethods(decision.allowMethods);
